Termination protection

+++ formal/resources/resource_policy.go
@@ -122,6 +122,13 @@ func ResourcePolicy() *schema.Resource {
 					Type: schema.TypeString,
 				},
 			},
+			"termination_protection": {
+				// This description is used by the documentation generator and the language server.
+				Description: "If set to true, this Policy cannot be deleted.",
+				Type:        schema.TypeBool,
+				Optional:    true,
+				Default:     false,
+			},
 		},
 	}
 }
@@ -145,16 +152,18 @@ func resourcePolicyCreate(ctx context.Context, d *schema.ResourceData, meta inte
 	Notification := d.Get("notification").(string)
 	Active := d.Get("active").(bool)
 	Status := d.Get("status").(string)
+	TerminationProtection := d.Get("termination_protection").(bool)
 
 	res, err := c.Grpc.Sdk.PolicyServiceClient.CreatePolicyV2(ctx, connect.NewRequest(&adminv1.CreatePolicyV2Request{
-		Name:         Name,
-		Description:  Description,
-		Code:         Module,
-		Notification: Notification,
-		Owners:       owners,
-		SourceType:   SourceType,
-		Active:       Active,
-		Status:       Status,
+		Name:                  Name,
+		Description:           Description,
+		Code:                  Module,
+		Notification:          Notification,
+		Owners:                owners,
+		SourceType:            SourceType,
+		Active:                Active,
+		Status:                Status,
+		TerminationProtection: TerminationProtection,
 	}))
 	if err != nil {
 		return diag.FromErr(err)
@@ -192,6 +201,7 @@ func resourcePolicyRead(ctx context.Context, d *schema.ResourceData, meta interf
 	d.Set("owners", res.Msg.Policy.Owners)
 	d.Set("active", res.Msg.Policy.Active)
 	d.Set("status", res.Msg.Policy.Status)
+	d.Set("termination_protection", res.Msg.Policy.TerminationProtection)
 
 	d.SetId(policyId)
 
@@ -203,7 +213,7 @@ func resourcePolicyUpdate(ctx context.Context, d *schema.ResourceData, meta inte
 
 	policyId := d.Id()
 
-	if d.HasChange("name") || d.HasChange("description") || d.HasChange("module") || d.HasChange("notification") || d.HasChange("owners") || d.HasChange("active") || d.HasChange("status") {
+	if d.HasChange("name") || d.HasChange("description") || d.HasChange("module") || d.HasChange("notification") || d.HasChange("owners") || d.HasChange("active") || d.HasChange("status") || d.HasChange("termination_protection") {
 
 		var owners []string
 		for _, owner := range d.Get("owners").([]interface{}) {
@@ -217,17 +227,19 @@ func resourcePolicyUpdate(ctx context.Context, d *schema.ResourceData, meta inte
 		SourceType := "terraform"
 		Active := d.Get("active").(bool)
 		Status := d.Get("status").(string)
+		TerminationProtection := d.Get("termination_protection").(bool)
 
 		_, err := c.Grpc.Sdk.PolicyServiceClient.UpdatePolicyV2(ctx, connect.NewRequest(&adminv1.UpdatePolicyV2Request{
-			Id:           policyId,
-			SourceType:   SourceType,
-			Name:         Name,
-			Description:  Description,
-			Code:         Module,
-			Notification: Notification,
-			Owners:       owners,
-			Active:       Active,
-			Status:       Status,
+			Id:                    policyId,
+			SourceType:            SourceType,
+			Name:                  Name,
+			Description:           Description,
+			Code:                  Module,
+			Notification:          Notification,
+			Owners:                owners,
+			Active:                Active,
+			Status:                Status,
+			TerminationProtection: TerminationProtection,
 		}))
 
 		if err != nil {
@@ -246,6 +258,12 @@ func resourcePolicyDelete(ctx context.Context, d *schema.ResourceData, meta inte
 	var diags diag.Diagnostics
 
 	policyId := d.Id()
+
+	terminationProtection := d.Get("termination_protection").(bool)
+
+	if terminationProtection {
+		return diag.Errorf("Policy cannot be deleted because termination_protection is set to true")
+	}
 
 	_, err := c.Grpc.Sdk.PolicyServiceClient.DeletePolicy(ctx, connect.NewRequest(&adminv1.DeletePolicyRequest{Id: policyId}))
 	if err != nil {
